requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.2
jax==0.4.25
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.10
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pip==20.2.3
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==49.2.1
six==1.16.0
sounddevice==0.4.6
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.23
urllib3==2.2.1
zipp==3.17.0
